added first working draft for portfolio page

+++ app/src/main/java/com/funglejunk/stockz/data/fboerse/FBoerseHistoryData.kt
@@ -19,7 +19,7 @@ data class FBoerseHistoryData(
         val close: Double,
         val high: Double,
         val low: Double,
-        val turnoverPieces: Int,
+        val turnoverPieces: Double, // TODO this seems not to be an int
         val turnoverEuro: Double
     )
 }
